+++ p5-server/src/server/liveReload.ts
@@ -6,16 +6,16 @@ export const liveReloadTemplate = `<script>
   ':$(port)/livereload.js?snipver=1"></' + 'script>')
 </script>`;
 
-export function injectLiveReloadScript(content: string, liveReloadServer: WebSocket.Server) {
+export function injectLiveReloadScript(html: string, liveReloadServer: WebSocket.Server) {
   // TODO: more robust injection
   // TODO: warn when injection is not possible
-  if (!liveReloadServer) return content;
+  if (!liveReloadServer) return html;
   const address = liveReloadServer.address();
   if (typeof address === 'string') {
     throw new Error('liveReloadServer.address is a string, not a WebSocket.AddressInfo');
   }
   const liveReloadString = liveReloadTemplate.replace('$(port)', address.port.toString());
-  return content.replace(/(?=<\/head>)/, liveReloadString);
+  return html.replace(/(?=<\/head>)/, liveReloadString);
 }
 
 export function createLiveReloadServer(watchDirs: string[]) {